backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
psycopg[binary]>=3.1,<4
django-environ>=0.11,<1
dj-database-url>=2.1,<3
celery>=5.4,<6
redis>=5.0,<6
